backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
SQLAlchemy>=1.4.0
pymysql>=1.0.2
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.0
